Collapsed forwarding support in ATS core
This adds a new write fail action, CACHE_WL_FAIL_ACTION_COLLAPSED_FORWARDING. If it fails to obtain a write lock then it will loop for open_write_retries*open_read_retry_time, each time doing a DecideCacheLookup.  In the end if it fails to ever either obtain a
write lock or find the object to read then it does the default write fail action of going to the origin.

In testing using duplicated requests I can see multiple simultaneous requests end up in this read loop and eventually get a hit in cache while the original request gets written.  Occaisonally there is still an extra hit or two to the origin which I believe is
state machine dependant, so sort of unavoidable, sometimes you can have the original request finish the write while a duplicated request ends up obtaining the write lock directly afterwards and so it hits the origin as well. This may be unavoidable unless cache
lookups are done everywhere before a write.

On an origin fail it will do the default of sending back an ERR_CONNECT_FAIL

(cherry picked from commit 72471ed52406752217576618b21a5901d336c9cf)

Add remove_ims to all collapsed forwarding states.  Currently we only remove the ims once we have run out of retries, so that it acts like the default case. The WL_READ_RETRY case also removes the ims so we might need it every time we loop from a write back to read?

Collapsed forwarding support in ATS core

This adds a new write fail action, CACHE_WL_FAIL_ACTION_COLLAPSED_FORWARDING. If it fails to obtain a write lock then it will loop for open_write_retries*open_read_retry_time, each time doing a DecideCacheLookup.  In the end if it fails to ever either obtain a
write lock or find the object to read then it does the default write fail action of going to the origin.

In testing using duplicated requests I can see multiple simultaneous requests end up in this read loop and eventually get a hit in cache while the original request gets written.  Occaisonally there is still an extra hit or two to the origin which I believe is
state machine dependant, so sort of unavoidable, sometimes you can have the original request finish the write while a duplicated request ends up obtaining the write lock directly afterwards and so it hits the origin as well. This may be unavoidable unless cache
lookups are done everywhere before a write.

On an origin fail it will do the default of sending back an ERR_CONNECT_FAIL

(cherry picked from commit 72471ed52406752217576618b21a5901d336c9cf)

Add remove_ims to all collapsed forwarding states.  Currently we only remove the ims once we have run out of retries, so that it acts like the default case. The WL_READ_RETRY case also removes the ims so we might need it every time we loop from a write back to read?

commenting out assert, I dont think this assert is necessary since the values are immediately overwritten

First shot a crash fix, needs a lot more pounding to see if its good. This adds a check of the cache_resp_hdr_heap_handle->m_heap->m_magic value to see if the actual data is good. It looks like when we get the crash the assigned *handle value is non-null but the data in there is garbage so it later fails the buffer checks. This allocates a new buffer if the handle is null or if the magic value is junk

Change to always use the cached_hdr heap

Revert changes to this file, the fix has been pushed to master and 81x

Change debug for CF so that only the first attempt is logged, and then the max if we hit it. This may help come up with more usable stats

Remove debug line that always prints when under retry amount

Collapsed forwarding support in ATS core

This adds a new write fail action, CACHE_WL_FAIL_ACTION_COLLAPSED_FORWARDING. If it fails to obtain a write lock then it will loop for open_write_retries*open_read_retry_time, each time doing a DecideCacheLookup.  In the end if it fails to ever either obtain a
write lock or find the object to read then it does the default write fail action of going to the origin.

In testing using duplicated requests I can see multiple simultaneous requests end up in this read loop and eventually get a hit in cache while the original request gets written.  Occaisonally there is still an extra hit or two to the origin which I believe is
state machine dependant, so sort of unavoidable, sometimes you can have the original request finish the write while a duplicated request ends up obtaining the write lock directly afterwards and so it hits the origin as well. This may be unavoidable unless cache
lookups are done everywhere before a write.

On an origin fail it will do the default of sending back an ERR_CONNECT_FAIL

(cherry picked from commit 72471ed52406752217576618b21a5901d336c9cf)

Add remove_ims to all collapsed forwarding states.  Currently we only remove the ims once we have run out of retries, so that it acts like the default case. The WL_READ_RETRY case also removes the ims so we might need it every time we loop from a write back to read?

Collapsed forwarding support in ATS core

This adds a new write fail action, CACHE_WL_FAIL_ACTION_COLLAPSED_FORWARDING. If it fails to obtain a write lock then it will loop for open_write_retries*open_read_retry_time, each time doing a DecideCacheLookup.  In the end if it fails to ever either obtain a
write lock or find the object to read then it does the default write fail action of going to the origin.

In testing using duplicated requests I can see multiple simultaneous requests end up in this read loop and eventually get a hit in cache while the original request gets written.  Occaisonally there is still an extra hit or two to the origin which I believe is
state machine dependant, so sort of unavoidable, sometimes you can have the original request finish the write while a duplicated request ends up obtaining the write lock directly afterwards and so it hits the origin as well. This may be unavoidable unless cache
lookups are done everywhere before a write.

On an origin fail it will do the default of sending back an ERR_CONNECT_FAIL

(cherry picked from commit 72471ed52406752217576618b21a5901d336c9cf)

Add remove_ims to all collapsed forwarding states.  Currently we only remove the ims once we have run out of retries, so that it acts like the default case. The WL_READ_RETRY case also removes the ims so we might need it every time we loop from a write back to read?

commenting out assert, I dont think this assert is necessary since the values are immediately overwritten

First shot a crash fix, needs a lot more pounding to see if its good. This adds a check of the cache_resp_hdr_heap_handle->m_heap->m_magic value to see if the actual data is good. It looks like when we get the crash the assigned *handle value is non-null but the data in there is garbage so it later fails the buffer checks. This allocates a new buffer if the handle is null or if the magic value is junk

Change to always use the cached_hdr heap

Revert changes to this file, the fix has been pushed to master and 81x

Change debug for CF so that only the first attempt is logged, and then the max if we hit it. This may help come up with more usable stats

Remove debug line that always prints when under retry amount

Remove func calls to get retry mode and #, not necessary

fix debug

Cancel pending action before we start doing any cache lookup. I dont believe there should be any actions at this point anyway since cache lookup is the first thing we do, however if we have looped around from a failed CF write then it is possible there is an action sitting there, so this helps to clean that up

Change debug for collapsed forwarding to use the cf_debug tag so it can be found separately from other logs

Remove cancel of pending action. We had a crash without having CF turned on so it is not the cause and its also not valid to be cancelling actions here