Add maxmind acl plugin
This plugin is similar to the current geoip acl plugin however it is based on the libmaxminddb library.  The GeoIP library is considered legacy at this point and it only supports the old dat formatted databases.
This will work with the newer more common mmdb format.

It takes a yaml formatted configuration file as shown in the docs, which specifies the database to use for that remap, and then any sets of allow and deny rules.  Currently the only rules supported are `country`,`ip`, and `regex`.
Deny rules will always take precedence if both allow and deny rules are set, and in the case of having both sets of rules or only allow rules the default action is to deny a connection. However if no accept rules are
specified then the default action will flip to always allow, since it doesnt make much sense to leave it as blocking everything if you only want to deny.  In that case all connections are allowed except those that fall into
one of the deny rules.