Add database reload on either IP lookup error, or mmdb lookup failures
We can run into a scenario where because the database is mmapped, if the file is changed underneath the mmap the results are unknown. The mmap docs state this especially if the size has changed which will happen
on mmdb updates. So we must reload the database in that instance. This is already handled by doing a remap reload, however there can be a time period between doing a touch/reload and when the db is updated
that users will just see 403's. This initiates an automatic reload of the database if we get any sort of lookup failures against it, once. After that an error is returned anyway

Still thinking about this though since Im not sure if under heavy traffic we could get in to a situation with multiple transactions all attempting a reload due to failures